+++ common-libs/common/src/_defaultConfig.ts
@@ -1,6 +1,3 @@
-export const DEFAULT_REGISTRY_URL = 'https://affinity-registry.staging.affinity-project.org'
-export const DEFAULT_METRICS_URL = 'https://affinity-metrics.staging.affinity-project.org'
-
 export const JOLO_DID_METHOD = 'jolo'
 export const ELEM_DID_METHOD = 'elem'
 export const ELEM_ANCHORED_DID_METHOD = 'elem-anchored'