+++ common-libs/url-resolver/.eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+  extends: ["@affinidi/eslint-config"],
+  parserOptions: {
+    project: './tsconfig.json'
+  },
+  env: {
+    "commonjs": true,
+    "node": true,
+    "mocha": true
+  }
+};